refactor: inserted pinch-controller into zoom-canvas.controller
