+++ packages/mapo-webapp/src/app/services/zoom-canvas/zoom-canvas.controller.ts
@@ -1,13 +1,17 @@
 import { Injectable } from '@angular/core';
 import { EventEmitter } from '@angular/core';
+import { fabric } from 'fabric';
 import { CanvasService } from '../canvas/canvas.service';
-import { ZoomCanvasService } from '../mouse-wheel/zoom-canvas.service';
+import { ZoomCanvasService } from './zoom-canvas.service';
 import Hammer from 'hammerjs';
 
+/**
+ * When the canvas is active, listens for mouse-scroll events, and zooms the canvas in or out.
+ */
 @Injectable({
   providedIn: 'root',
 })
-export class PinchController {
+export class MouseWheelController {
   public pinchStateChange: EventEmitter<boolean> = new EventEmitter();
   lastScale = 1;
   constructor(
@@ -40,9 +44,23 @@ export class PinchController {
           this.onPinch(event, canvas);
         });
       }
+      canvas.on('mouse:wheel', this.onMouseWheel);
     });
-    this.canvasService.canvasDestroyed$.subscribe((canvas) => {});
+    this.canvasService.canvasDestroyed$.subscribe((canvas) => {
+      canvas.off('mouse:wheel', this.onMouseWheel as any);
+    });
   }
+
+  onMouseWheel = (event: fabric.IEvent<WheelEvent>) => {
+    const delta = event.e.deltaY;
+    const x = event.e.offsetX;
+    const y = event.e.offsetY;
+
+    this.zoomCanvasService.zoomCanvas(delta, x, y);
+
+    event.e.preventDefault();
+    event.e.stopPropagation();
+  };
 
   onPinch = (event: HammerInput, canvas: fabric.Canvas) => {
     const x = event.center.x;
