+++ packages/mapo-webapp/src/app/services/zoom-canvas/zoom-canvas.service.ts
